+++ packages/errors/src/ZeroDepGuardError.ts
@@ -1,5 +1,7 @@
-export class ZeroDepError extends Error {
-  constructor(message = 'An unexpected error has occurred') {
+import { ZeroDepError } from './ZeroDepError';
+
+export class ZeroDepGuardError extends ZeroDepError {
+  constructor(message = 'Guard Error') {
     super(message);
 
     // ensure the name matches the class