+++ packages/errors/src/index.spec.ts
@@ -2,16 +2,20 @@ import * as packages from './index';
 
 const keys = Object.keys(packages).sort();
 
-describe('Errors barrel package', () => {
+describe('Errors package', () => {
   it('should export specific packages and interfaces', () => {
-    expect(keys).toStrictEqual(['ZeroDepError']);
+    expect(keys).toStrictEqual([
+      'ZeroDepError',
+      'ZeroDepGuardError',
+      'ZeroDepTypeError',
+    ]);
   });
 
   // contrived test for barrel-file
   test.each(keys)('should have %s', (name) => {
     expect(['function', 'object', 'undefined']).toContain(
       // @ts-ignore
-      typeof packages[name]
+      typeof packages[name],
     );
   });
 });