+++ packages/guard/function/src/lib/guardFunction.ts
@@ -7,7 +7,6 @@ export const guardFunctionHOF = () => {
     // we need to check for the typeof first as "undefined" will cause isFunction() to error
     if (!isFunction(value)) {
       const error = new ZeroDepError('Value is not a function');
-      error.value = value;
       throw error;
     }
   };