+++ packages/guard/integer/src/lib/guardInteger.ts
@@ -19,19 +19,16 @@ export const guardIntegerHOF = (options: GuardIntegerOptions = {}) => {
     // we need to check for the typeof first as "undefined" will cause isInteger() to error
     if (!isInteger(value)) {
       const error = new ZeroDepError('Value is not an integer');
-      error.value = value;
       throw error;
     }
 
     if (typeof config.min !== 'undefined' && (value as number) < config.min) {
       const error = new ZeroDepError(`Integer is less than ${config.min}`);
-      error.value = value;
       throw error;
     }
 
     if (typeof config.max !== 'undefined' && (value as number) > config.max) {
       const error = new ZeroDepError(`Integer is greater than ${config.max}`);
-      error.value = value;
       throw error;
     }
   };